requirements.txt
cloudshell-core==2.2.176
git+https://github.com/QualiSystems/cloudshell-L1-networking-core.git@dev#egg=cloudshell_l1_networking_core
git+https://github.com/QualiSystems/cloudshell-cli.git@dev#egg=cloudshell-cli
